app/requirements.txt
flask
turbo-flask
pyzmq
distinctipy
redis